Consolidate ConversionCompletedView with VC
